chore(team): split mailbox channel tests
